add getDistance method to ICellOM ceil strength value
