+++ .gitignore
@@ -39,3 +39,4 @@ build/
 
 ### Mac OS ###
 .DS_Store
+/.idea/
