--- src/main/java/it/rcs/om/ICellOM.java
@@ -1,7 +1,0 @@
-package it.rcs.om;
-
-public interface ICellOM {
-
-    double getStrength(PositionOM position);
-
-}
