--- src/main/java/it/rcs/om/PositionOM.java
@@ -1,11 +1,0 @@
-package it.rcs.om;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class PositionOM {
-    private double latitude;
-    private double longitude;
-}
