+++ src/main/java/it/rcs/om/cell/ICellOM.java
@@ -1,0 +1,7 @@
+package it.rcs.om.cell;
+
+public interface ICellOM {
+
+    double getStrength(PositionOM position);
+    double getDistance(PositionOM startingPoint, PositionOM endingPoint);
+}
