+++ src/main/java/it/rcs/om/cell/RadiusCellOM.java
@@ -1,4 +1,4 @@
-package it.rcs.om;
+package it.rcs.om.cell;
 
 import lombok.Getter;
 import lombok.Setter;
@@ -16,7 +16,7 @@ public class RadiusCellOM extends AbstractCellOM {
 
     @Override
     public double getStrength(PositionOM pointPosition) {
-        double distance = getDistance(pointPosition);
-        return Math.max(0, 100 - (distance / radius * 100));
+        double distance = getDistance(pointPosition, this.getPosition());
+        return Math.ceil(Math.max(0, 100 - (distance / radius * 100)));
     }
 }
