+++ src/main/java/it/rcs/om/event/EventTypeOM.java
@@ -1,4 +1,4 @@
-package it.rcs.om;
+package it.rcs.om.event;
 
 public enum EventTypeOM {
     CONNECTED,
